requirements.txt
fastapi
uvicorn
pandas
joblib
requests
python-dotenv
scikit-learn